pyproject.toml
[tool.poetry]
name = "CSEAN Demo"
version = "0.1.0"
description = "Event invitation management demo"
authors = ["its-ohio <ohio.imevbore@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
pyairtable = "^2.3.3"
flask = "^3.0.3"
requests = "^2.32.3"
jupyter = "^1.0.0"
streamlit = "^1.38.0"
streamlit-js-eval = "^0.1.7"
yagmail = "^0.15.293"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
